elytra: Clean kernel from out directory on installclean
KERNEL_OBJ directory was just a symlink to this directory.
Hence, on installclean, KERNEL_OBJ was getting wiped while
kernel directory was still there in out. On running the build,
this symlink gets recreated. Since it matches the previous
condition of the directory, the build doesn't consider
rebuilding the kernel.

• This commit does address the issue at [1].

Reference:
[1] https://gerrit.aospa.co/c/AOSPA/android_build_soong/+/26101

Change-Id: Idbd998de19f2a634cc9d9c93ccb921bb00d41718